UIDATIMP-1403: Rename mutatorMock and okapiMock
